Fix camera plane and better timer handling
Former-commit-id: 80d61e0c3878144dad15f0b7495332ea2846dee3 [formerly 53dbbe51624bfc7813867da94dd723b618d2a804]
Former-commit-id: 6049c9cabe09ef1bc9239a0197c0bcf81f872fc0
Former-commit-id: 3b1b34fa777db82651396f6d923ead9b0a0c1829
Former-commit-id: a9984f7cb9570315ee4a0a2094f7adcea5ec0add

+++ src/frontends/Qt/canvas.cpp
@@ -28,14 +28,13 @@ void Canvas::paintEvent(QPaintEvent *event) {
     #ifndef WIREFRAME
       p.setPen(Qt::NoPen);
       p.setBrush(QColor(
-        triangle.color.x(),
-        triangle.color.y(),
-        triangle.color.z()
+        static_cast<int>(triangle.color.x()),
+        static_cast<int>(triangle.color.y()),
+        static_cast<int>(triangle.color.z())
       ));
     #endif
     p.drawConvexPolygon(points, 3);
-  }  
-
+  }
 }
 
 void Canvas::update_frame(Rect b) {
@@ -52,8 +51,8 @@ QPointF Canvas::adjust_coordinates(const Point2& p) {
 }
 
 void Canvas::resizeEvent(QResizeEvent *event) {
-  x_offset = event->size().width()  / 2;
-  y_offset = event->size().height() / 2;
+  x_offset = static_cast<double>(event->size().width())  / 2;
+  y_offset = static_cast<double>(event->size().height()) / 2;
 }
 
 void Canvas::set_triangles_buffer(const std::vector<Triangle2> *buff_a,
